MA-24280 Fix system cannot boot up normally when not connect display bridge
Find new situation that no connector but crtc and plane are initialized normally,
need to regard this case as no any physical display and dummy display will be used.

Find that the resolution of dummy display should be large enough to match with
system navigationbar related setting, or the launcher3 or systemui will crash
sometimes. Change the default resolution of dummy display to 1080p.

Fix some build warning because format mismatch when ALOGI.

Change-Id: I37ded091f9af999c9e33b41c72a7d6999fcd6e95
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/30343
Reviewed-by: Jessie Hao <juan.hao@nxp.com>