MA-24095 audio: Recover saved configs when stop
A resampler is created when the input stream could
not be opened with the requested rate.
The mConfig->rate is updated to 48K in ::start() and
the resampler is released later in ::stop().
If the stream wants to start again, the mConfig
has to be recovered to the original value, otherwise
the resampler will not be created anymore.

Test: MediaRecorderTest#testSetCamera
Change-Id: I84a4f8dd7fd9186ed40962b068fac658cbaeaaef
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/30563
Reviewed-by: Hui Fang <hui.fang@nxp.com>