MA-24149-11 avoid affect performance when not enable g2d optimization
The performance drop a little on the platforms that without g2d optimization.
It's because the clearRect() operation is not necessary for all cases. Use
clearWormHole() instead of clearRect() to avoid unnecessary clear target operation.
Refine the logic of clearWormHole() to ignore small opaque UI region, it means
to clear such region to avoid cursor mess when video playback with overlay.

Change-Id: I9d7bb2a6865529c864de2ccde83bdf5fe03ef802
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/30618
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>