MA-24149-12 Avoid black block on UI when there are multiple visible region in layer
Find that when in desktop window mode, playback two video with Gallery and
CactusPlayer at the same time, then drag the CactusPlayer window to overlap
the Gallery app, there are some black blocks on UI sometimes.
It's because the wallpaper layer has multiple visible region in such case,
also there are some unvisible region on it. When it composed as intermediate
composition buffer, the unvisible region will keep while the other layer move
away and cannot cover the unvisible region. So we can see the black block on
UI sometimes.
Compose the whole layer to inter compostion buffer to avoid such issue.

Change-Id: I65dd1708d41f89990329c4517cbfd2db20bf2141
Signed-off-by: Bo Zhang <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/30619
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>